Add files via upload
This script parse gmap coordinates into the syntenic block where the pan-gene is coming from. The output of this file includes a list of the coordinates of the genes that need to be added into the pan gene matrix. It also outputs a gff file where it could be intersected  with the existing annotation gff file.